pets_site/pyproject.toml
[tool.poetry]
name = "pets"
version = "0.1.0"
description = ""
authors = ["NikSen <niksen09@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djoser = "^2.2.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
pillow = "^10.4.0"
drf-extra-fields = "^3.7.0"
gunicorn = "^22.0.0"
django-cleanup = "^8.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
